pyproject.toml
[project]
name = "continual"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.9.21"
dependencies = [
    "avalanche-lib>=0.6.0",
    "matplotlib>=3.9.4",
    "numpy>=2.0.2",
    "pandas>=2.2.3",
    "ray[tune]>=2.43.0",
    "seaborn>=0.13.2",
    "sparse>=0.15.5",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
]
